+++ src/editors/Configuration/store/ConfigurationStore.test.js
@@ -1,6 +1,6 @@
 import Store from "./ConfigurationStore";
-import DBSubscriber from "../DBSubscriber";
-import Model from "../models/Configuration";
+import { DBSubscriber } from "../../../store";
+import { Model } from "../../../models";
 
 test("Smoke test", () => {
   const obj = new Store();